working on db
