+++ server/index.js
@@ -2,7 +2,7 @@
 
 require('dotenv').config();
 console.log(process.env)
-const { start } = require('./server.js');
+const { start } = require('./src/server.js');
 const PORT = process.env.PORT
 console.log(PORT);
